# backend/internal/httpapi/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API do LogN. Erro de rota que o app lê sai como {\"code\", \"message\"}, e o código é contrato.",
        "title": "LogN API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/nodes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Grafo de nós e cartões de origem",
                "parameters": [
                    {
                        "enum": [
                            "pt-BR",
                            "en",
                            "es"
                        ],
                        "type": "string",
                        "description": "Língua do conteúdo",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpapi.nodesResponse"
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    }
                }
            }
        },
        "/api/v1/progress": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Progresso da conta",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserStats"
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    },
                    "500": {
                        "description": "internal",
                        "schema": {
                            "$ref": "#/definitions/httpapi.apiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.OriginCard": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "domain.SkillNode": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prerequisites": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "required_xp": {
                    "type": "integer"
                },
                "requires_purchase": {
                    "description": "O nó é de trilha paga e não é a amostra: só abre com licença. Quem decide é o\nservidor, pela mesma regra que filtra os desafios abertos (`openNode`).",
                    "type": "boolean"
                },
                "row": {
                    "type": "integer"
                },
                "topic": {
                    "description": "Assunto do nó, neutro (`adhoc`, `graphs`): decide cor e ícone no app, que antes\nadivinhava pelo nome. Vazio enquanto o conteúdo não preencher.",
                    "type": "string"
                },
                "track_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.UserProgress": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "current_xp": {
                    "type": "integer"
                },
                "node_id": {
                    "type": "string"
                },
                "unlocked": {
                    "type": "boolean"
                }
            }
        },
        "domain.UserStats": {
            "type": "object",
            "properties": {
                "bugs_found": {
                    "type": "integer"
                },
                "dry_runs_completed": {
                    "type": "integer"
                },
                "global_xp": {
                    "type": "integer"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UserProgress"
                    }
                },
                "paid_challenge_ids": {
                    "description": "Desafios que já renderam XP. O cliente precisa deles para não prometer XP de novo\npor um desafio que o servidor não vai pagar, inclusive depois de trocar de aparelho.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "paid_track_xp": {
                    "description": "XP ganho em cada trilha paga, amostra incluída. O portão de um nó conta só o XP da\ntrilha dele (spec, seção 8); o da trilha gratuita é o global menos a soma destes.\n\nSai daqui e não do Core porque o Core pode não conhecer mais os desafios de uma\ntrilha revogada, e aí somaria o XP dela no portão da gratuita.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "httpapi.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "httpapi.nodesResponse": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SkillNode"
                    }
                },
                "origins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.OriginCard"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token do login, no formato `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}